merge

+++ litemall-wx-api/src/main/java/org/linlinjava/litemall/wx/web/WxAddressController.java
@@ -133,20 +133,28 @@ public class WxAddressController extends GetRegionService {
 			return error;
 		}
 
-		if (address.getIsDefault()) {
-			// 重置其他收货地址的默认选项
-			addressService.resetDefault(userId);
-		}
-
 		if (address.getId() == null || address.getId().equals(0)) {
+			if (address.getIsDefault()) {
+				// 重置其他收货地址的默认选项
+				addressService.resetDefault(userId);
+			}
+
 			address.setId(null);
 			address.setUserId(userId);
 			addressService.add(address);
 		} else {
-			address.setUserId(userId);
-			if (addressService.update(address) == 0) {
-				return ResponseUtil.updatedDataFailed();
+			LitemallAddress litemallAddress = addressService.query(userId, address.getId());
+			if (litemallAddress == null) {
+				return ResponseUtil.badArgumentValue();
 			}
+
+			if (address.getIsDefault()) {
+				// 重置其他收货地址的默认选项
+				addressService.resetDefault(userId);
+			}
+
+			address.setUserId(userId);
+			addressService.update(address);
 		}
 		return ResponseUtil.ok(address.getId());
 	}
@@ -166,6 +174,10 @@ public class WxAddressController extends GetRegionService {
 		Integer id = address.getId();
 		if (id == null) {
 			return ResponseUtil.badArgument();
+		}
+		LitemallAddress litemallAddress = addressService.query(userId, id);
+		if (litemallAddress == null) {
+			return ResponseUtil.badArgumentValue();
 		}
 
 		addressService.delete(id);
